Decoupled publication and subscription from main controller file.
